images/upload-gitlab-failure-logs/requirements.txt
opensearch-dsl==2.0.1
python-gitlab==3.11.0
PyYAML==6.0
psycopg2-binary==2.9.5
